Refactor: 커스텀훅으로 반응형, 폼 제출 로직 추상화
