+++ src/components/Layout/Header.tsx
@@ -1,39 +1,14 @@
-import { ChangeEvent, FormEvent, useEffect, useState } from "react";
 import { useRouter } from "next/router";
 import { SearchInput } from "@/components/SearchInput";
 import Logo from "../Logo";
 import Navigation from "../Navigation";
 import HeaderMenuContainer from "../../containers/HeaderMenuContainer";
+import useViewport from "@/hooks/useViewport";
+import useSearchName from "../../hooks/useSearchName";
 
 export const Header = () => {
-  const router = useRouter();
-  const [isMobile, setIsMobile] = useState(false);
-  const [searchedName, setSearchedName] = useState("");
-
-  const handleResize = () => {
-    setIsMobile(window.innerWidth < 768);
-  };
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setSearchedName(e.target.value);
-  };
-
-  const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    router.push({
-      pathname: "/search",
-      query: { q: searchedName },
-    });
-    setSearchedName("");
-  };
-
-  useEffect(() => {
-    window.addEventListener("resize", handleResize);
-    handleResize();
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
+  const { searchedName, handleChange, handleSubmit } = useSearchName();
+  const { isMobile } = useViewport();
 
   return (
     <div className="w-screen p-[24px] h-[80px] bg-background flex justify-between items-center shadow-sm sticky top-0 z-[90]">
